server/requirements.txt
Flask
flask-cors
Pillow
openai
python-dotenv
requests
schedule
